deleted RedhawkGCInterface class
