+++ src/coreclr/nativeaot/Runtime/CMakeLists.txt
@@ -23,7 +23,7 @@ set(COMMON_RUNTIME_SOURCES
     ObjectLayout.cpp
     portable.cpp
     RestrictedCallouts.cpp
-    RedhawkGCInterface.cpp
+    GcEnum.cpp
     RhConfig.cpp
     RuntimeInstance.cpp
     StackFrameIterator.cpp
