+++ src/coreclr/nativeaot/Runtime/gcenv.ee.cpp
@@ -11,7 +11,7 @@
 #include "RestrictedCallouts.h"
 
 #include "RhConfig.h"
-#include "RedhawkGCInterface.h"
+#include "GcEnum.h"
 
 #include "thread.h"
 #include "threadstore.h"
@@ -115,12 +115,12 @@ void GCToEEInterface::GcScanRoots(ScanFunc* fn, int condemned, int max_gen, Scan
             while (pRoot != NULL)
             {
                 STRESS_LOG2(LF_GC | LF_GCROOTS, LL_INFO100, "{ Scanning Thread's %p inline thread statics root %p. \n", pThread, pRoot);
-                RedhawkGCInterface::EnumGcRef(&pRoot->m_threadStaticsBase, GCRK_Object, fn, sc);
+                EnumGcRef(&pRoot->m_threadStaticsBase, GCRK_Object, fn, sc);
                 pRoot = pRoot->m_next;
             }
 
             STRESS_LOG1(LF_GC | LF_GCROOTS, LL_INFO100, "{ Scanning Thread's %p thread statics root. \n", pThread);
-            RedhawkGCInterface::EnumGcRef(pThread->GetThreadStaticStorage(), GCRK_Object, fn, sc);
+            EnumGcRef(pThread->GetThreadStaticStorage(), GCRK_Object, fn, sc);
 
             STRESS_LOG1(LF_GC | LF_GCROOTS, LL_INFO100, "{ Starting scan of Thread %p\n", pThread);
             sc->thread_under_crawl = pThread;
